fixed two warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.3@4549 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/MATH/MISC/NonNegativeLeastSquaresSolver.C
@@ -60,8 +60,9 @@ namespace OpenMS
 		//std::cout << "A:\n" << A << std::endl;
 		#endif
 		
-		Int a_rows = A.rows();
-		Int a_cols = A.cols();
+		// this needs to be int (not Int, Size or anything else), because the external nnls constructor expects it this way!
+		int a_rows = (int)A.rows();
+		int a_cols = (int)A.cols();
 		
 		// translate b
 		double *b_vec = new double[a_rows];
